U.Codebase.Sqlite.Branch.Format

Documentation

data BranchFormat' text defRef patchRef childRef branchRef

A BranchFormat is a deserialized namespace object (object.bytes).

you can use the exact same BranchLocalIds when converting between Full and BranchFormat'

Constructors

Full (BranchLocalIds' text defRef patchRef childRef) LocalBranch

Diff branchRef (BranchLocalIds' text defRef patchRef childRef) LocalDiff

type BranchFormat = BranchFormat' TextId ObjectId PatchObjectId (BranchObjectId, CausalHashId) BranchObjectId

The BranchFormat' used to store a branch in Sqlite

type HashBranchFormat = BranchFormat' Text ComponentHash PatchHash (BranchHash, CausalHash)

A BranchFormat which uses Hashes and Text for all its references, no Ids which are specific to a particular codebase.

type BranchLocalIds = BranchLocalIds' TextId ObjectId PatchObjectId (BranchObjectId, CausalHashId)

A BranchLocalIds is a mapping between local ids (local to this object) encoded as offsets, and actual database ids.

For example, a branchTextLookup vector of [50, 74] means "local id 0 corresponds to database text id 50, and local id 1 corresponds to database text id 74".

data BranchLocalIds' t d p c

Constructors

LocalIds

Fields

branchTextLookup :: Vector t

branchDefnLookup :: Vector d

branchPatchLookup :: Vector p

branchChildLookup :: Vector c

type HashBranchLocalIds = BranchLocalIds' Text ComponentHash PatchHash (BranchHash, CausalHash)

type SyncBranchFormat = SyncBranchFormat' BranchObjectId TextId ObjectId PatchObjectId (BranchObjectId, CausalHashId)

data SyncBranchFormat' parent text defn patch child

Constructors

SyncFull (BranchLocalIds' text defn patch child) LocalBranchBytes

SyncDiff parent (BranchLocalIds' text defn patch child) LocalBranchBytes

newtype LocalBranchBytes

Bytes encoding a LocalBranch

Constructors

LocalBranchBytes ByteString

localToDbBranch :: BranchLocalIds -> LocalBranch -> DbBranch

localToDbDiff :: BranchLocalIds -> LocalDiff -> Diff

localToHashBranch :: HashBranchLocalIds -> LocalBranch -> HashBranch

localToBranch :: (Ord t, Ord d) => BranchLocalIds' t d p c -> LocalBranch -> Branch' t d p c