U.Codebase.Sqlite.Decode

Description

This module contains decoders for blobs stored in SQLite.

Documentation

data DecodeError

Constructors

DecodeError

Fields

decoder :: Text

err :: String

object.bytes

decodeBranchFormat :: ByteString -> Either DecodeError BranchFormat

decodeComponentLengthOnly :: ByteString -> Either DecodeError Word64

decodeDeclElement :: Word64 -> ByteString -> Either DecodeError (LocalIds, Decl Symbol)

decodeDeclElementNumConstructors :: Word64 -> ByteString -> Either DecodeError Int

decodeDeclFormat :: ByteString -> Either DecodeError DeclFormat

decodePatchFormat :: ByteString -> Either DecodeError PatchFormat

decodeSyncDeclFormat :: ByteString -> Either DecodeError SyncDeclFormat

decodeSyncNamespaceFormat :: ByteString -> Either DecodeError SyncBranchFormat

decodeSyncPatchFormat :: ByteString -> Either DecodeError SyncPatchFormat

decodeSyncTermFormat :: ByteString -> Either DecodeError SyncTermFormat

decodeSyncTermAndType :: ByteString -> Either DecodeError (Term, Type)

N.B. The bytestring here is not the entire object.bytes column -- it's just the serialized term and type from SyncTermFormat.

decodeTermElementDiscardingTerm :: Pos -> ByteString -> Either DecodeError (LocalIds, Type)

decodeTermElementDiscardingType :: Pos -> ByteString -> Either DecodeError (LocalIds, Term)

decodeTermElementWithType :: Pos -> ByteString -> Either DecodeError (LocalIds, Term, Type)

decodeTermFormat :: ByteString -> Either DecodeError TermFormat

temp_entity.blob

decodeTempCausalFormat :: ByteString -> Either DecodeError TempCausalFormat

decodeTempDeclFormat :: ByteString -> Either DecodeError TempDeclFormat

decodeTempNamespaceFormat :: ByteString -> Either DecodeError TempNamespaceFormat

decodeTempPatchFormat :: ByteString -> Either DecodeError TempPatchFormat

decodeTempTermFormat :: ByteString -> Either DecodeError TempTermFormat

watch_result.result

decodeWatchResultFormat :: ByteString -> Either DecodeError WatchResultFormat

unsyncs

unsyncTermComponent :: HasCallStack => SyncLocallyIndexedComponent' t d -> Either DecodeError (LocallyIndexedComponent' t d)

unsyncDeclComponent :: SyncLocallyIndexedComponent' t d -> Either DecodeError (LocallyIndexedComponent' t d)