U.Codebase.Sqlite.HashHandle

Documentation

data HashHandle

Constructors

HashHandle

Fields

toReference :: Type Symbol -> Reference

Hash type

toReferenceMentions :: Type Symbol -> Set Reference

Hash type's mentions

toReferenceDecl :: Hash -> TypeD Symbol -> Reference

Hash the type of a single constructor in a decl component. The provided hash argument is the hash of the decl component.

toReferenceDeclMentions :: Hash -> TypeD Symbol -> Set Reference

Hash decl's mentions