U.Codebase.Sqlite.LocalIds

Documentation

data LocalIds' t h

A mapping between index ids that are local to an object and the ids in the database

Constructors

LocalIds

Fields

textLookup :: Vector t

defnLookup :: Vector h

type LocalIds = LocalIds' TextId ObjectId

type WatchLocalIds = LocalIds' TextId HashId

newtype LocalTextId

represents an index into a textLookup

Constructors

LocalTextId Word64

newtype LocalDefnId

represents an index into a defnLookup

In this context, "definition" means an object that is either a term component or a (type) decl component, not a patch, namespace, or any other kind of object.

Constructors

LocalDefnId Word64

newtype LocalHashId

a local index to a hash, used when the corresponding object is allowed to be absent

Constructors

LocalHashId Word64

newtype LocalPatchObjectId

Constructors

LocalPatchObjectId Word64

newtype LocalBranchChildId

Constructors

LocalBranchChildId Word64

newtype LocalCausalHashId

causal hashes are treated differently from HashIds, which don't have dependencies

Constructors

LocalCausalHashId Word64

t_ :: Traversal (LocalIds' t h) (LocalIds' t' h) t t'

h_ :: Traversal (LocalIds' t h) (LocalIds' t h') h h'