U.Codebase.Sqlite.NameLookups

Description

Types related to working with NameLookups. We define these low-level types rather than use Path's because we don't have access to those domain types given the package dependency tree.

Documentation

newtype ReversedName

Constructors

ReversedName (NonEmpty Text)

newtype ReversedPath

Constructors

ReversedPath [Text]

newtype PathSegments

Constructors

PathSegments [Text]

type NamespaceText = Text

A namespace rendered as a path, no leading . E.g. "base.data"

reversedNameToNamespaceText :: ReversedName -> NamespaceText

>>> reversedSegmentsToNamespaceText (["List", "data", "base"])
"base.data.List"

reversedNameToPathSegments :: ReversedName -> PathSegments

pathSegmentsToText :: PathSegments -> Text

>>> pathSegmentsToText (PathSegments ["base", "data", "List"])
"base.data.List"

textToPathSegments :: Text -> PathSegments

>>> textToPathSegments "base.data.List"
PathSegments ["base","data","List"]