U.Codebase.Sqlite.NamedRef

Documentation

data ConstructorType

Constructors

DataConstructor

EffectConstructor

data NamedRef ref

Constructors

NamedRef

Fields

reversedSegments :: ReversedName

ref :: ref

newtype ScopedRow ref

The new scoped name lookup format is different from the old version.

Specifically, the scoped format adds the lastNameSegment as well as adding a trailing . to the db format of both the namespace and reversed_name.

This type has a ToRow instance of the form: [reversedName, namespace, lastNameSegment] <> ref fields...

Constructors

ScopedRow (NamedRef ref)