U.Codebase.Sqlite.Operations

branches

loadCausalHashAtPath :: CausalHash -> [NameSegment] -> Transaction (Maybe CausalHash)

Load the causal hash at the given path from the provided root, if Nothing, use the codebase root.

expectCausalHashAtPath :: CausalHash -> [NameSegment] -> Transaction CausalHash

Expect the causal hash at the given path from the provided root, if Nothing, use the codebase root.

loadCausalBranchAtPath :: CausalHash -> [NameSegment] -> Transaction (Maybe (CausalBranch Transaction))

loadBranchAtPath :: CausalHash -> [NameSegment] -> Transaction (Maybe (Branch Transaction))

saveBranch :: HashHandle -> CausalBranch Transaction -> Transaction (BranchObjectId, CausalHashId)

saveBranchV3 :: HashHandle -> CausalBranchV3 Transaction -> Transaction (BranchObjectId, CausalHashId)

loadCausalBranchByCausalHash :: CausalHash -> Transaction (Maybe (CausalBranch Transaction))

expectCausalBranchByCausalHash :: CausalHash -> Transaction (CausalBranch Transaction)

expectBranchByCausalHashId :: CausalHashId -> Transaction (Branch Transaction)

expectBranchByBranchHash :: BranchHash -> Transaction (Branch Transaction)

expectBranchByBranchHashId :: BranchHashId -> Transaction (Branch Transaction)

expectNamespaceStatsByHash :: BranchHash -> Transaction NamespaceStats

Looks up statistics for a given branch, if none exist, we compute them and save them then return them.

expectNamespaceStatsByHashId :: BranchHashId -> Transaction NamespaceStats

Looks up statistics for a given branch, if none exist, we compute them and save them then return them.

tryGetSquashResult :: BranchHash -> Transaction (Maybe CausalHash)

Get the causal hash which would be the result of squashing the provided branch hash. Returns Nothing if we haven't computed it before.

saveSquashResult :: BranchHash -> CausalHash -> Transaction ()

Saves the result of a squash

terms

saveTermComponent

Arguments

:: HashHandle

-> Maybe ByteString
The serialized term component if we already have it e.g. via sync

-> Hash
term component hash

-> [(Term Symbol, Type Symbol)]
term component

-> Transaction ObjectId

loadTermComponent :: Hash -> MaybeT Transaction [(Term Symbol, Type Symbol)]

loadTermByReference :: Id -> MaybeT Transaction (Term Symbol)

loadTypeOfTermByTermReference :: Id -> MaybeT Transaction (Type Symbol)

decls

saveDeclComponent :: HashHandle -> Maybe ByteString -> Hash -> [Decl Symbol] -> Transaction ObjectId

loadDeclComponent :: Hash -> MaybeT Transaction [Decl Symbol]

loadDeclByReference :: Id -> MaybeT Transaction (Decl Symbol)

expectDeclByReference :: Id -> Transaction (Decl Symbol)

expectDeclNumConstructors :: Id -> Transaction Int

expectDeclTypeById :: Id -> Transaction DeclType

Get the DeclType of a Id.

terms/decls

getCycleLen :: Hash -> Transaction (Maybe Word64)

patches

savePatch :: HashHandle -> PatchHash -> Patch -> Transaction PatchObjectId

expectPatch :: PatchObjectId -> Transaction Patch

test for stuff in codebase

objectExistsForHash :: Hash -> Transaction Bool

watch expression cache

saveWatch :: WatchKind -> Id -> Term Symbol -> Transaction ()

loadWatch :: WatchKind -> Id -> MaybeT Transaction (Term Symbol)

returns Nothing if the expression isn't cached.

listWatches :: WatchKind -> Transaction [Id]

clearWatches :: Transaction ()

Delete all watches that were put by putWatch.

indexes

nearest common ancestor

before :: CausalHash -> CausalHash -> Transaction (Maybe Bool)

lca :: CausalHash -> CausalHash -> Transaction (Maybe CausalHash)

prefix index

componentReferencesByPrefix :: ObjectType -> Text -> Maybe Pos -> Transaction [Id]

termReferentsByPrefix :: Text -> Maybe Word64 -> Transaction [Id]

declReferentsByPrefix :: Text -> Maybe Pos -> Maybe ConstructorId -> Transaction [(Hash, Pos, DeclType, [ConstructorId])]

causalHashesByPrefix :: ShortCausalHash -> Transaction (Set CausalHash)

dependents index

directDependenciesOfScope :: (Reference -> Bool) -> DefnsF Set TermReferenceId TypeReferenceId -> Transaction (DefnsF Set TermReference TypeReference)

dependents :: DependentsSelector -> Reference -> Transaction (Set Id)

returns a list of known definitions referencing r

dependentsOfComponent :: Hash -> Transaction (Set Id)

returns a list of known definitions referencing h

directDependentsWithinScope :: DefnsF Set TermReferenceId TypeReferenceId -> DefnsF Set TermReference TypeReference -> Transaction (DefnsF Set TermReferenceId TypeReferenceId)

`directDependentsWithinScope scope query` returns all direct dependents of query that are in scope (excluding self-references).

transitiveDependentsWithinScope :: DefnsF Set TermReferenceId TypeReferenceId -> DefnsF Set TermReference TypeReference -> Transaction (DefnsF Set TermReferenceId TypeReferenceId)

`transitiveDependentsWithinScope scope query` returns all transitive dependents of query that are in scope (excluding self-references).

data DependencyEdge

Constructors

TermDependsOnTerm TermReferenceId TermReference

TermDependsOnType TermReferenceId TypeReference

TypeDependsOnType TypeReferenceId TypeReference

transitiveDependentsGraphWithinScope :: (Reference -> Bool) -> DefnsF Set TermReferenceId TypeReferenceId -> DefnsF Set TermReference TypeReference -> Transaction [DependencyEdge]

type index

addTypeToIndexForTerm :: Id -> Reference -> Transaction ()

termsHavingType :: Reference -> Transaction (Set Id)

filterTermsByReferenceHavingType :: TypeReference -> [Id] -> Transaction [Id]

filterTermsByReferentHavingType :: TypeReference -> [Id] -> Transaction [Id]

type mentions index

addTypeMentionsToIndexForTerm :: Id -> Set Reference -> Transaction ()

termsMentioningType :: Reference -> Transaction (Set Id)

Projects

expectProjectAndBranchNames :: ProjectId -> ProjectBranchId -> Transaction (ProjectName, ProjectBranchName)

expectProjectBranchHead :: ProjectId -> ProjectBranchId -> Transaction CausalHash

reflog

getDeprecatedRootReflog :: Int -> Transaction [Entry CausalHash Text]

Gets the specified number of reflog entries in chronological order, most recent first.

getProjectReflog :: Int -> ProjectId -> Transaction [Entry Project ProjectBranch CausalHash]

Gets the specified number of reflog entries for the given project in chronological order, most recent first.

getProjectBranchReflog :: Int -> ProjectBranchId -> Transaction [Entry Project ProjectBranch CausalHash]

Gets the specified number of reflog entries for the specified ProjectBranch in chronological order, most recent first.

getGlobalReflog :: Int -> Transaction [Entry Project ProjectBranch CausalHash]

Gets the specified number of reflog entries in chronological order, most recent first.

appendProjectReflog :: Entry ProjectId ProjectBranchId CausalHash -> Transaction ()

low-level stuff

expectDbBranch :: BranchObjectId -> Transaction DbBranch

saveDbBranch :: HashHandle -> BranchHash -> NamespaceStats -> DbBranchV -> Transaction BranchObjectId

Save a DbBranch, given its hash (which the caller is expected to produce from the branch).

Note: long-standing question: should this package depend on the hashing package? (If so, we would only need to take the DbBranch, and hash internally).

saveDbBranchUnderHashId :: HashHandle -> BranchHashId -> NamespaceStats -> DbBranchV -> Transaction BranchObjectId

Variant of saveDbBranch that might be preferred by callers that already have a hash id, not a hash.

expectDbPatch :: PatchObjectId -> Transaction Patch

saveDbPatch :: HashHandle -> PatchHash -> PatchFormat -> Transaction PatchObjectId

expectDbBranchByCausalHashId :: CausalHashId -> Transaction DbBranch

Expect a branch value given its causal hash id.

namespaceStatsForDbBranch :: DbBranchV -> Transaction NamespaceStats

reference conversions

c2sTextReference :: Reference -> TextReference

s2cTextReference :: TextReference -> Reference

c2sTextReferent :: Referent -> TextReferent

s2cTextReferent :: TextReferent -> Referent

somewhat unexpectedly unused definitions

c2sReferenceId :: Id -> Transaction Id

c2sReferentId :: Id -> Transaction Id

diffPatch :: LocalPatch -> LocalPatch -> LocalPatchDiff

produces a diff diff = full - ref; full = diff + ref

decodeTermElementWithType :: Pos -> ByteString -> Either DecodeError (LocalIds, Term, Type)

loadTermWithTypeByReference :: Id -> MaybeT Transaction (Term Symbol, Type Symbol)

s2cTermWithType :: LocalIds -> Term -> Type -> Transaction (Term Symbol, Type Symbol)

s2cDecl :: LocalIds -> Decl Symbol -> Transaction (Decl Symbol)

Unlocalize a decl.

declReferencesByPrefix :: Text -> Maybe Word64 -> Transaction [Id]

namespaceHashesByPrefix :: ShortNamespaceHash -> Transaction (Set BranchHash)

derivedDependencies :: Id -> Transaction (Set Id)

returns empty set for unknown inputs; doesn't distinguish between term and decl

internal stuff that probably need not be exported, but the 1->2 migration needs it

data BranchV (m :: Type -> Type)

Constructors

BranchV2 !(Branch m)

BranchV3 !(BranchV3 m)

data DbBranchV

Constructors

DbBranchV2 !DbBranch

DbBranchV3 !DbBranchV3