U.Codebase.Sqlite.Patch.Format

Documentation

data PatchFormat

Constructors

Full PatchLocalIds LocalPatch

Diff PatchObjectId PatchLocalIds LocalPatchDiff

type PatchLocalIds = PatchLocalIds' TextId HashId ObjectId

data PatchLocalIds' t h d

Constructors

LocalIds

Fields

patchTextLookup :: Vector t

patchHashLookup :: Vector h

patchDefnLookup :: Vector d

type HashPatchLocalIds = PatchLocalIds' Text ComponentHash ComponentHash

LocalIds type which can be used in hashing the Patch.

type SyncPatchFormat = SyncPatchFormat' PatchObjectId TextId HashId ObjectId

data SyncPatchFormat' parent text hash defn

Constructors

SyncFull (PatchLocalIds' text hash defn) ByteString

SyncDiff parent (PatchLocalIds' text hash defn) ByteString
p is the identity of the thing that the diff is relative to

applyPatchDiffs :: Patch -> [PatchDiff] -> Patch

Apply a list of patch diffs to a patch, left to right.

localPatchToPatch :: PatchLocalIds -> LocalPatch -> Patch

Type specialized version of localToPatch'.

localPatchToPatch' :: (Ord t, Ord h, Ord d) => PatchLocalIds' t h d -> Patch' LocalTextId LocalHashId LocalDefnId -> Patch' t h d

Generic version of localPatchToPatch that works with any PatchLocalIds'.

localPatchDiffToPatchDiff :: PatchLocalIds -> LocalPatchDiff -> PatchDiff

localPatchToHashPatch :: HashPatchLocalIds -> LocalPatch -> HashPatch