U.Codebase.Sqlite.ProjectReflog

Documentation

data Entry project branch causal

Constructors

Entry

Fields

project :: project

branch :: branch

time :: UTCTime

fromRootCausalHash :: Maybe causal

toRootCausalHash :: causal

reason :: Text

project_ :: Lens (Entry project branch causal) (Entry project' branch causal) project project'

branch_ :: Lens (Entry project branch causal) (Entry project branch' causal) branch branch'

projectAndBranch_ :: Lens (Entry project branch causal) (Entry project' branch' causal) (project, branch) (project', branch')

Both Project and Branch Ids are required to load a branch, so this is often more useful.