U.Codebase.Sqlite.Queries

Description

Some naming conventions used in this module:

32: the base32 representation of a hash

expect: retrieve something that's known to exist

load: retrieve something that's not known to exist (so the return type is a Maybe, or another container that could be empty)

save: idempotent (on conflict do nothing) insert, and return the id of the thing (usually)

text table

saveText :: Text -> Transaction TextId

saveTexts :: Traversable f => f Text -> Transaction (f TextId)

loadTextId :: Text -> Transaction (Maybe TextId)

expectTextId :: Text -> Transaction TextId

expectText :: TextId -> Transaction Text

expectTextCheck :: SqliteExceptionReason e => TextId -> (Text -> Either e a) -> Transaction a

hash table

saveHash :: Hash32 -> Transaction HashId

saveHashes :: Traversable f => f Hash32 -> Transaction (f HashId)

saveHashHash :: Hash -> Transaction HashId

loadHashId :: Hash32 -> Transaction (Maybe HashId)

expectHash :: HashId -> Transaction Hash

expectHash32 :: HashId -> Transaction Hash32

expectBranchHash :: BranchHashId -> Transaction BranchHash

loadHashIdByHash :: Hash -> Transaction (Maybe HashId)

expectHashIdByHash :: Hash -> Transaction HashId

saveCausalHash :: CausalHash -> Transaction CausalHashId

expectCausalHash :: CausalHashId -> Transaction CausalHash

saveBranchHash :: BranchHash -> Transaction BranchHashId

hash_object table

saveHashObject :: HashId -> ObjectId -> HashVersion -> Transaction ()

expectHashIdsForObject :: ObjectId -> Transaction (NonEmpty HashId)

hashIdWithVersionForObject :: ObjectId -> Transaction [(HashId, HashVersion)]

loadObjectIdForPrimaryHashId :: HashId -> Transaction (Maybe ObjectId)

expectObjectIdForPrimaryHashId :: HashId -> Transaction ObjectId

Not all hashes have corresponding objects; e.g., hashes of term types

loadObjectIdForPrimaryHash :: Hash -> Transaction (Maybe ObjectId)

expectObjectIdForPrimaryHash :: Hash -> Transaction ObjectId

loadPatchObjectIdForPrimaryHash :: PatchHash -> Transaction (Maybe PatchObjectId)

loadObjectIdForAnyHash :: Hash -> Transaction (Maybe ObjectId)

loadObjectIdForAnyHashId :: HashId -> Transaction (Maybe ObjectId)

expectObjectIdForAnyHashId :: HashId -> Transaction ObjectId

recordObjectRehash :: ObjectId -> ObjectId -> Transaction ()

recordObjectRehash old new records that object old was rehashed and inserted as a new object, new.

This function rewrites old's hash_object rows in place to point at the new object.

object table

saveObject :: HashHandle -> HashId -> ObjectType -> ByteString -> Transaction ObjectId

isObjectHash :: HashId -> Transaction Bool

Does a hash correspond to an object?

expectObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

expectPrimaryHashByObjectId :: ObjectId -> Transaction Hash

All objects have corresponding hashes.

expectPrimaryHashIdForObject :: ObjectId -> Transaction HashId

expectObjectWithHashIdAndType :: ObjectId -> Transaction (HashId, ObjectType, ByteString)

expectDeclObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a decl component object.

loadDeclObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a decl component object.

expectNamespaceObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a namespace object.

loadNamespaceObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a namespace object.

expectPatchObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a patch object.

loadPatchObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a patch object.

loadTermObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a term component object.

expectTermObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a term component object.

namespace_root table

loadNamespaceRoot :: Transaction (Maybe CausalHashId)

setNamespaceRoot :: CausalHashId -> Transaction ()

expectNamespaceRoot :: Transaction CausalHashId

causals

causal table

saveCausal :: HashHandle -> CausalHashId -> BranchHashId -> [CausalHashId] -> Transaction ()

Maybe we would generalize this to something other than NamespaceHash if we end up wanting to store other kinds of Causals here too.

isCausalHash :: HashId -> Transaction Bool

expectCausal :: CausalHashId -> Transaction SyncCausalFormat

loadCausalHashIdByCausalHash :: CausalHash -> Transaction (Maybe CausalHashId)

expectCausalHashIdByCausalHash :: CausalHash -> Transaction CausalHashId

expectCausalValueHashId :: CausalHashId -> Transaction BranchHashId

loadCausalByCausalHash :: CausalHash -> Transaction (Maybe (CausalHashId, BranchHashId))

expectCausalByCausalHash :: CausalHash -> Transaction (CausalHashId, BranchHashId)

loadBranchObjectIdByCausalHashId :: CausalHashId -> Transaction (Maybe BranchObjectId)

expectBranchObjectIdByCausalHashId :: CausalHashId -> Transaction BranchObjectId

causal_parent table

saveCausalParents :: CausalHashId -> [CausalHashId] -> Transaction ()

loadCausalParents :: CausalHashId -> Transaction [CausalHashId]

loadCausalParentsByHash :: Hash32 -> Transaction [Hash32]

Like loadCausalParents, but the input and outputs are hashes, not hash ids.

before :: CausalHashId -> CausalHashId -> Transaction Bool

lca :: CausalHashId -> CausalHashId -> Transaction (Maybe CausalHashId)

watch table

saveWatch :: WatchKind -> IdH -> ByteString -> Transaction ()

loadWatch :: SqliteExceptionReason e => WatchKind -> IdH -> (ByteString -> Either e a) -> Transaction (Maybe a)

loadWatchesByWatchKind :: WatchKind -> Transaction [IdH]

loadWatchKindsByReference :: IdH -> Transaction [WatchKind]

clearWatches :: Transaction ()

indexes

dependents index

addToDependentsIndex :: Reference -> Id -> Transaction ()

getDependentsForDependency :: Reference -> Transaction [Id]

Get non-self, user-defined dependents of a dependency.

getDependentsForDependencyComponent :: ObjectId -> Transaction [Id]

getDependenciesForDependent :: Id -> Transaction [Reference]

Get non-self dependencies of a user-defined dependent.

getDependencyIdsForDependent :: Id -> Transaction [Id]

Get non-self, user-defined dependencies of a user-defined dependent.

migrations

countObjects :: Transaction Int

countCausals :: Transaction Int

countWatches :: Transaction Int

getCausalsWithoutBranchObjects :: Transaction [CausalHashId]

Finds all causals that refer to a branch for which we don't have an object stored. Although there are plans to support this in the future, currently all such cases are the result of database inconsistencies and are unexpected.

removeHashObjectsByHashingVersion :: HashVersion -> Transaction ()

Delete all hash objects of a given hash version. Leaves the corresponding hashes in the hash table alone.

type index

addToTypeIndex :: Reference' TextId HashId -> Id -> Transaction ()

getReferentsByType :: Reference' TextId HashId -> Transaction [Id]

getTypeReferenceForReferent :: Id -> Transaction (Reference' TextId HashId)

getTypeReferencesForComponent :: ObjectId -> Transaction [(Reference' TextId HashId, Id)]

type mentions index

addToTypeMentionsIndex :: Reference' TextId HashId -> Id -> Transaction ()

getReferentsByTypeMention :: Reference' TextId HashId -> Transaction [Id]

getTypeMentionsReferencesForComponent :: ObjectId -> Transaction [(Reference' TextId HashId, Id)]

hash prefix lookup

objectIdByBase32Prefix :: ObjectType -> Text -> Transaction [ObjectId]

namespaceHashIdByBase32Prefix :: Text -> Transaction [BranchHashId]

causalHashIdByBase32Prefix :: Text -> Transaction [CausalHashId]

Name Lookup

resetNameLookupTables :: Transaction ()

Drop and recreate the name lookup tables. Use this when resetting names to a new branch.

insertTermNames :: [NamedRef (TextReferent, Maybe ConstructorType)] -> Transaction ()

Insert the given set of term names into the name lookup table

insertTypeNames :: [NamedRef TextReference] -> Transaction ()

Insert the given set of type names into the name lookup table

rootTermNames :: Transaction [NamedRef (TextReferent, Maybe ConstructorType)]

Get the list of a term names in the root namespace according to the name lookup index

rootTypeNames :: Transaction [NamedRef TextReference]

Get the list of a type names in the root namespace according to the name lookup index

garbage collection

garbageCollectObjectsWithoutHashes :: Transaction ()

Delete objects without hashes. An object typically *would* have a hash, but (for example) during a migration in which an object's hash may change, its corresponding hash_object row may be updated to point at a new version of that object. This procedure clears out all references to objects that do not have any corresponding hash_object rows.

garbageCollectWatchesWithoutObjects :: Transaction ()

Delete all

sync temp entities

data EntityLocation

Where an entity is stored.

Constructors

EntityInMainStorage
object / causal

EntityInTempStorage
temp_entity

entityExists :: Hash32 -> Transaction Bool

Does this entity already exist in the database, i.e. in the object or causal table?

entityLocation :: Hash32 -> Transaction (Maybe EntityLocation)

Where is an entity stored?

expectEntity :: Hash32 -> Transaction SyncEntity

Read an entity out of main storage.

syncToTempEntity :: SyncEntity -> Transaction TempEntity

looking up all of the text and hashes is the first step of converting a SyncEntity to a Share.Entity

insertTempEntity :: Hash32 -> TempEntity -> NEMap Hash32 Text -> Transaction ()

Insert a new temp_entity row, and its associated 1+ temp_entity_missing_dependency rows.

Preconditions: 1. The entity does not already exist in "main" storage (object / causal) 2. The entity does not already exist in temp_entity.

saveTempEntityInMain :: HashHandle -> Hash32 -> TempEntity -> Transaction (Either CausalHashId ObjectId)

Save a temp entity in main storage.

Precondition: all of its dependencies are already in main storage.

expectTempEntity :: Hash32 -> Transaction TempEntity

Read an entity out of temp storage.

deleteTempEntity :: Hash32 -> Transaction ()

Delete a row from the temp_entity table, if it exists.

elaborate hashes

elaborateHashes :: NonEmpty Hash32 -> Transaction (NonEmpty Text)

Elaborate a set of temp_entity hashes.

Given a set of temp_entity hashes, returns the (known) set of transitive dependencies that haven't already been downloaded (i.e. aren't in the temp_entity table)

For example, if we have temp entities A and B, where A depends on B and B depends on C...

| temp_entity | | temp_entity_missing_dependency | |=============| |================================| | hash | | dependent | dependency | |-------------| |--------------|-----------------| | A | | A | B | | B | | B | C |

... then `elaborateHashes {A}` would return the singleton set {C} (because we take the set of transitive dependencies {A,B,C} and subtract the set we already have, {A,B}).

db misc

addTempEntityTables :: Transaction ()

addTypeMentionsToIndexForTerm :: Id -> Set Reference -> Transaction ()

addTypeToIndexForTerm :: Id -> Reference -> Transaction ()

c2xTerm :: forall m t d. Monad m => (Text -> m t) -> (Hash -> m d) -> Term Symbol -> Maybe (Type Symbol) -> m (LocalIds' t d, Term, Maybe Type)

implementation detail of c2{s,w}Term The Type is optional, because we don't store them for watch expression results.

createSchema :: Transaction ()

expectSchemaVersion :: SchemaVersion -> Transaction ()

Expect the given schema version.

localIdsToLookups :: Monad m => (t -> m Text) -> (d -> m Hash) -> LocalIds' t d -> m (LocalTextId -> Text, LocalDefnId -> Hash)

implementation detail of {s,w}2c*Term* & s2cDecl

s2cDecl :: LocalIds -> Decl Symbol -> Transaction (Decl Symbol)

Unlocalize a decl.

s2cTermWithType :: LocalIds -> Term -> Type -> Transaction (Term Symbol, Type Symbol)

saveDeclComponent :: HashHandle -> Maybe ByteString -> Hash -> [Decl Symbol] -> Transaction ObjectId

saveReferenceH :: Reference -> Transaction ReferenceH

Save the text and hash parts of a Reference to the database and substitute their ids.

saveSyncEntity :: HashHandle -> Hash32 -> SyncEntity -> Transaction (Either CausalHashId ObjectId)

saveTermComponent

Arguments

:: HashHandle

-> Maybe ByteString
The serialized term component if we already have it e.g. via sync

-> Hash
term component hash

-> [(Term Symbol, Type Symbol)]
term component

-> Transaction ObjectId

schemaVersion :: Transaction SchemaVersion

setSchemaVersion :: SchemaVersion -> Transaction ()

x2cTType :: (LocalTextId -> Text) -> (LocalDefnId -> Hash) -> Type -> Type Symbol

implementation detail of {s,w}2c*Term*

x2cTerm :: (LocalTextId -> Text) -> (LocalDefnId -> Hash) -> Term -> Term Symbol

implementation detail of {s,w}2c*Term*