U.Codebase.Sqlite.Serialization

Documentation

debug :: Bool

putABT :: (MonadPut m, Foldable f, Functor f, Ord v) => (v -> m ()) -> (a -> m ()) -> (forall x. (x -> m ()) -> f x -> m ()) -> Term f v a -> m ()

getABT :: (MonadGet m, Foldable f, Functor f, Ord v) => m v -> m a -> (forall x. m x -> m (f x)) -> m (Term f v a)

putLocalIds :: (MonadPut m, Integral t, Bits t, Integral d, Bits d) => LocalIds' t d -> m ()

getLocalIds :: MonadGet m => m LocalIds

getWatchLocalIds :: MonadGet m => m WatchLocalIds

putUnit :: Applicative m => () -> m ()

getUnit :: Applicative m => m ()

putWatchResultFormat :: MonadPut m => WatchResultFormat -> m ()

getWatchResultFormat :: MonadGet m => m WatchResultFormat

putTermFormat :: MonadPut m => TermFormat -> m ()

getTermFormat :: MonadGet m => m TermFormat

putTermComponent :: MonadPut m => LocallyIndexedComponent -> m ()

putTerm :: MonadPut m => Term -> m ()

getTermComponent :: MonadGet m => m LocallyIndexedComponent

getTerm :: MonadGet m => m Term

lookupTermElement :: MonadGet m => Pos -> m (LocalIds, Term, Type)

lookupTermElementDiscardingType :: MonadGet m => Pos -> m (LocalIds, Term)

lookupTermElementDiscardingTerm :: MonadGet m => Pos -> m (LocalIds, Type)

getTType :: MonadGet m => m Type

getType :: MonadGet m => m r -> m (TypeR r Symbol)

putDeclFormat :: MonadPut m => DeclFormat -> m ()

getDeclFormat :: MonadGet m => m DeclFormat

getDeclElement :: MonadGet m => m (Decl Symbol)

lookupDeclElement :: MonadGet m => Pos -> m (LocalIds, Decl Symbol)

putBranchFormat :: MonadPut m => BranchFormat -> m ()

putBranchLocalIds :: MonadPut m => BranchLocalIds -> m ()

putPatchFormat :: MonadPut m => PatchFormat -> m ()

getPatchFormat :: MonadGet m => m PatchFormat

getPatchLocalIds :: MonadGet m => m PatchLocalIds

putPatchFull :: MonadPut m => LocalPatch -> m ()

putPatchDiff :: MonadPut m => LocalPatchDiff -> m ()

putPatchLocalIds :: MonadPut m => PatchLocalIds -> m ()

putTermEdit :: MonadPut m => LocalTermEdit -> m ()

putTypeEdit :: MonadPut m => LocalTypeEdit -> m ()

getBranchFormat :: MonadGet m => m BranchFormat

getBranchLocalIds :: MonadGet m => m BranchLocalIds

decomposeComponent :: MonadGet m => m [(LocalIds, ByteString)]

recomposeComponent :: MonadPut m => [(LocalIds, ByteString)] -> m ()

decomposeWatchFormat :: MonadGet m => m SyncWatchResultFormat

recomposeWatchFormat :: MonadPut m => SyncWatchResultFormat -> m ()

decomposePatchFormat :: MonadGet m => m SyncPatchFormat

recomposePatchFormat :: MonadPut m => SyncPatchFormat -> m ()

decomposeBranchFormat :: MonadGet m => m SyncBranchFormat

recomposeBranchFormat :: MonadPut m => SyncBranchFormat -> m ()

getSymbol :: MonadGet m => m Symbol

putSymbol :: MonadPut m => Symbol -> m ()

putReferent :: (MonadPut m, Integral t1, Integral h1, Bits t1, Bits h1, Integral t2, Integral h2, Bits t2, Bits h2) => Referent' (Reference' t1 h1) (Reference' t2 h2) -> m ()

putReferent' :: MonadPut m => (r1 -> m ()) -> (r2 -> m ()) -> Referent' r1 r2 -> m ()

putReference :: (MonadPut m, Integral t, Bits t, Integral r, Bits r) => Reference' t r -> m ()

getReferent' :: MonadGet m => m r1 -> m r2 -> m (Referent' r1 r2)

getReferent :: (MonadGet m, Integral t1, Integral h1, Bits t1, Bits h1, Integral t2, Integral h2, Bits t2, Bits h2) => m (Referent' (Reference' t1 h1) (Reference' t2 h2))

getReference :: (MonadGet m, Integral t, Bits t, Integral r, Bits r) => m (Reference' t r)

putRecursiveReference :: (MonadPut m, Integral t, Bits t, Integral r, Bits r) => Reference' t (Maybe r) -> m ()

getRecursiveReference :: (MonadGet m, Integral t, Bits t, Integral r, Bits r) => m (Reference' t (Maybe r))

putInt :: MonadPut m => Int64 -> m ()

getInt :: MonadGet m => m Int64

putNat :: MonadPut m => Word64 -> m ()

getNat :: MonadGet m => m Word64

putFloat :: MonadPut m => Double -> m ()

getFloat :: MonadGet m => m Double

putBoolean :: MonadPut m => Bool -> m ()

getBoolean :: MonadGet m => m Bool

putTType :: MonadPut m => Type -> m ()

putDType :: MonadPut m => Type Symbol -> m ()

putType :: (MonadPut m, Ord v) => (r -> m ()) -> (v -> m ()) -> TypeR r v -> m ()

putChar :: MonadPut m => Char -> m ()

getChar :: MonadGet m => m Char

putMaybe :: MonadPut m => (a -> m ()) -> Maybe a -> m ()

getMaybe :: MonadGet m => m a -> m (Maybe a)

unknownTag :: (MonadGet m, Show a) => String -> a -> m x