U.Codebase.Sqlite.TempEntityType

Documentation

data TempEntityType

Don't reorder these, they are part of the database, and the ToField and FromField implementation currently depends on the derived Enum implementation.

Constructors

TermComponentType

DeclComponentType

NamespaceType

PatchType

CausalType