U.Codebase.Sqlite.Term.Format

Documentation

type TermRef = Reference' LocalTextId (Maybe LocalDefnId)

Builtin terms are represented as local text ids.

Non-builtin terms are represented as local definition ids, with an added distinguished element (here Nothing) which represents a self-reference.

type TypeRef = Reference' LocalTextId LocalDefnId

Builtin types are represented as a local text id.

Non-builtin types are represented by a local definition id.

type TermLink = Referent' TermRef TypeRef

type TypeLink = TypeRef

type LocallyIndexedComponent = LocallyIndexedComponent' TextId ObjectId

A LocallyIndexedComponent is a vector that has one element per member of the component (invariant: 1+).

Each element is a term, which is represented as:

Lookup vectors that map local ids to database ids for texts and objects referenced by the term.

The term itself, with internal references to local ids (offsets into the lookup vectors).

The term's type, also with internal references to local id.

type HashLocallyIndexedComponent = LocallyIndexedComponent' Text Hash32

A locally indexed component which uses hash references instead of database ids.

newtype LocallyIndexedComponent' t d

Constructors

LocallyIndexedComponent

Fields

unLocallyIndexedComponent :: Vector (LocalIds' t d, Term, Type)

newtype SyncLocallyIndexedComponent' t d

Constructors

SyncLocallyIndexedComponent (Vector (LocalIds' t d, ByteString))

type F = F' LocalTextId TermRef TypeRef TermLink TypeLink Symbol

type FT = F' TypeRef

type Term = Term F Symbol ()

type Type = Term FT Symbol ()

Type of Term

type FTT = F' Reference

type TypeOfTerm = Term FTT Symbol ()

type TermFormat = TermFormat' TextId ObjectId

type HashTermFormat = TermFormat' Text Hash32

A TermFormat which uses hash references instead of database ids.

data TermFormat' t d

Constructors

Term (LocallyIndexedComponent' t d)

type SyncTermFormat = SyncTermFormat' TextId ObjectId

data SyncTermFormat' t d

Constructors

SyncTerm (SyncLocallyIndexedComponent' t d)

data WatchResultFormat

Constructors

WatchResult WatchLocalIds Term

data SyncWatchResultFormat

Constructors

SyncWatchResult WatchLocalIds ByteString