U.Codebase.Sqlite.HashHandle

Documentation

data HashHandle

Constructors

HashHandle

Fields

toReference :: Type Symbol -> Reference

Hash type

toReferenceMentions :: Type Symbol -> Set Reference

Hash type's mentions

toReferenceDecl :: Hash -> TypeD Symbol -> Reference

Hash the type of a single constructor in a decl component. The provided hash argument is the hash of the decl component.

toReferenceDeclMentions :: Hash -> TypeD Symbol -> Set Reference

Hash decl's mentions

hashBranch :: forall m. Monad m => Branch m -> m BranchHash

hashBranchV3 :: forall m. BranchV3 m -> BranchHash

hashCausal :: BranchHash -> Set CausalHash -> CausalHash

hashBranchFormatFull :: HashBranchLocalIds -> LocalBranch -> BranchHash

hashPatchFormatFull :: HashPatchLocalIds -> LocalPatch -> PatchHash

verifyTermFormatHash :: ComponentHash -> HashTermFormat -> Maybe HashMismatch

verifyDeclFormatHash :: ComponentHash -> HashDeclFormat -> Maybe DeclHashingError

data HashMismatch

Constructors

HashMismatch

Fields

expectedHash :: Hash

actualHash :: Hash

data DeclHashingError

Constructors

DeclHashMismatch HashMismatch

DeclHashResolutionFailure