U.Codebase.Sqlite.Operations

branches

saveRootBranch :: HashHandle -> CausalBranch Transaction -> Transaction (BranchObjectId, CausalHashId)

loadRootCausalHash :: Transaction (Maybe CausalHash)

expectRootCausalHash :: Transaction CausalHash

expectRootCausal :: Transaction (CausalBranch Transaction)

loadCausalHashAtPath :: [Text] -> Transaction (Maybe CausalHash)

Load the causal hash at the given path from the root.

FIXME should we move some Path type here?

expectCausalHashAtPath :: [Text] -> Transaction CausalHash

Expect the causal hash at the given path from the root.

FIXME should we move some Path type here?

saveBranch :: HashHandle -> CausalBranch Transaction -> Transaction (BranchObjectId, CausalHashId)

loadCausalBranchByCausalHash :: CausalHash -> Transaction (Maybe (CausalBranch Transaction))

expectCausalBranchByCausalHash :: CausalHash -> Transaction (CausalBranch Transaction)

terms

saveTermComponent

Arguments

:: HashHandle

-> Maybe ByteString
The serialized term component if we already have it e.g. via sync

-> Hash
term component hash

-> [(Term Symbol, Type Symbol)]
term component

-> Transaction ObjectId

loadTermComponent :: Hash -> MaybeT Transaction [(Term Symbol, Type Symbol)]

loadTermByReference :: Id -> MaybeT Transaction (Term Symbol)

loadTypeOfTermByTermReference :: Id -> MaybeT Transaction (Type Symbol)

decls

saveDeclComponent :: HashHandle -> Maybe ByteString -> Hash -> [Decl Symbol] -> Transaction ObjectId

loadDeclComponent :: Hash -> MaybeT Transaction [Decl Symbol]

loadDeclByReference :: Id -> MaybeT Transaction (Decl Symbol)

expectDeclTypeById :: Id -> Transaction DeclType

Get the DeclType of a Id.

terms/decls

getCycleLen :: Hash -> Transaction (Maybe Word64)

patches

savePatch :: HashHandle -> PatchHash -> Patch -> Transaction PatchObjectId

expectPatch :: PatchObjectId -> Transaction Patch

test for stuff in codebase

objectExistsForHash :: Hash -> Transaction Bool

watch expression cache

saveWatch :: WatchKind -> Id -> Term Symbol -> Transaction ()

loadWatch :: WatchKind -> Id -> MaybeT Transaction (Term Symbol)

returns Nothing if the expression isn't cached.

listWatches :: WatchKind -> Transaction [Id]

clearWatches :: Transaction ()

indexes

nearest common ancestor

before :: CausalHash -> CausalHash -> Transaction (Maybe Bool)

lca :: CausalHash -> CausalHash -> Transaction (Maybe CausalHash)

prefix index

componentReferencesByPrefix :: ObjectType -> Text -> Maybe Pos -> Transaction [Id]

termReferentsByPrefix :: Text -> Maybe Word64 -> Transaction [Id]

declReferentsByPrefix :: Text -> Maybe Pos -> Maybe ConstructorId -> Transaction [(Hash, Pos, DeclType, [ConstructorId])]

causalHashesByPrefix :: ShortBranchHash -> Transaction (Set CausalHash)

dependents index

dependents :: Reference -> Transaction (Set Id)

returns a list of known definitions referencing r

dependentsOfComponent :: Hash -> Transaction (Set Id)

returns a list of known definitions referencing h

type index

addTypeToIndexForTerm :: Id -> Reference -> Transaction ()

termsHavingType :: Reference -> Transaction (Set Id)

type mentions index

addTypeMentionsToIndexForTerm :: Id -> Set Reference -> Transaction ()

termsMentioningType :: Reference -> Transaction (Set Id)

name lookup index

rebuildNameIndex :: [NamedRef (Referent, Maybe ConstructorType)] -> [NamedRef Reference] -> Transaction ()

Given the list of term and type names from the root branch, rebuild the name lookup table.

rootBranchNames :: Transaction ([NamedRef (Referent, Maybe ConstructorType)], [NamedRef Reference])

Get all the term and type names for the root namespace from the lookup table.

low-level stuff

expectDbBranch :: BranchObjectId -> Transaction DbBranch

saveDbBranch :: HashHandle -> BranchHash -> DbBranch -> Transaction BranchObjectId

Save a DbBranch, given its hash (which the caller is expected to produce from the branch).

Note: long-standing question: should this package depend on the hashing package? (If so, we would only need to take the DbBranch, and hash internally).

saveDbBranchUnderHashId :: HashHandle -> BranchHashId -> DbBranch -> Transaction BranchObjectId

Variant of saveDbBranch that might be preferred by callers that already have a hash id, not a hash.

expectDbPatch :: PatchObjectId -> Transaction Patch

saveDbPatch :: HashHandle -> PatchHash -> PatchFormat -> Transaction PatchObjectId

expectDbBranchByCausalHashId :: CausalHashId -> Transaction DbBranch

Expect a branch value given its causal hash id.

somewhat unexpectedly unused definitions

c2sReferenceId :: Id -> Transaction Id

c2sReferentId :: Id -> Transaction Id

diffPatch :: LocalPatch -> LocalPatch -> LocalPatchDiff

produces a diff diff = full - ref; full = diff + ref

decodeTermElementWithType :: Pos -> ByteString -> Either DecodeError (LocalIds, Term, Type)

loadTermWithTypeByReference :: Id -> MaybeT Transaction (Term Symbol, Type Symbol)

s2cTermWithType :: LocalIds -> Term -> Type -> Transaction (Term Symbol, Type Symbol)

s2cDecl :: LocalIds -> Decl Symbol -> Transaction (Decl Symbol)

Unlocalize a decl.

declReferencesByPrefix :: Text -> Maybe Word64 -> Transaction [Id]

branchHashesByPrefix :: ShortBranchHash -> Transaction (Set BranchHash)

derivedDependencies :: Id -> Transaction (Set Id)

returns empty set for unknown inputs; doesn't distinguish between term and decl