U.Codebase.Sqlite.Operations

Error handling

throwError :: Err m => Error -> m a

debug :: Bool

crashOnError :: Bool

crashOnError can be helpful for debugging. If it is False, the errors will be delivered to the user elsewhere.

type Err m = (MonadError Error m, HasCallStack)

type EDB m = (Err m, DB m)

type ErrString = String

data DecodeError

Constructors

ErrTermElement Word64

ErrDeclElement Word64

ErrFramedArrayLen

ErrTypeOfTerm Id

ErrWatch WatchKind Id

ErrBranch BranchObjectId

ErrPatch PatchObjectId

ErrObjectDependencies ObjectType ObjectId

data Error

Constructors

DecodeError DecodeError ByteString ErrString

DatabaseIntegrityError Integrity

UnknownDependency Hash

UnknownText Text

ExpectedBranch CausalHash BranchHash

ExpectedBranch' CausalHashId

LegacyUnknownCycleLen Hash

LegacyUnknownConstructorType Hash Pos

NeedTypeForBuiltinMetadata Text

getFromBytesOr :: Err m => DecodeError -> Get a -> ByteString -> m a

liftQ :: Err m => ExceptT Integrity m a -> m a

Database lookups

lookupTextId :: EDB m => Text -> m TextId

loadTextById :: EDB m => TextId -> m Text

primaryHashToExistingObjectId :: EDB m => Hash -> m ObjectId

look up an existing object by its primary hash

primaryHashToMaybeObjectId :: DB m => Hash -> m (Maybe ObjectId)

primaryHashToMaybePatchObjectId :: EDB m => PatchHash -> m (Maybe PatchObjectId)

objectExistsForHash :: DB m => Hash -> m Bool

loadHashByObjectId :: EDB m => ObjectId -> m Hash

loadHashByHashId :: EDB m => HashId -> m Hash

loadCausalHashById :: EDB m => CausalHashId -> m CausalHash

loadValueHashByCausalHashId :: EDB m => CausalHashId -> m BranchHash

loadRootCausalHash :: EDB m => m CausalHash

loadMaybeRootCausalHash :: EDB m => m (Maybe CausalHash)

Reference transformations

read existing references

c2sReference :: EDB m => Reference -> m Reference

s2cReference :: EDB m => Reference -> m Reference

c2sReferenceId :: EDB m => Id -> m Id

s2cReferenceId :: EDB m => Id -> m Id

h2cReferenceId :: EDB m => IdH -> m Id

h2cReference :: EDB m => ReferenceH -> m Reference

c2hReference :: DB m => Reference -> MaybeT m ReferenceH

s2cReferent :: EDB m => Referent -> m Referent

s2cReferentId :: EDB m => Id -> m Id

c2sReferentId :: EDB m => Id -> m Id

h2cReferent :: EDB m => ReferentH -> m Referent

Edits transformations

s2cTermEdit :: EDB m => TermEdit -> m TermEdit

s2cTyping :: Typing -> Typing

c2sTyping :: Typing -> Typing

s2cTypeEdit :: EDB m => TypeEdit -> m TypeEdit

c2lPatch :: EDB m => Patch -> m (PatchLocalIds, LocalPatch)

assumes that all relevant values are already in the DB

diffPatch :: LocalPatch -> LocalPatch -> LocalPatchDiff

produces a diff diff = full - ref; full = diff + ref

Deserialization helpers

decodeComponentLengthOnly :: Err m => ByteString -> m Word64

decodeTermElementWithType :: Err m => Pos -> ByteString -> m (LocalIds, Term, Type)

decodeTermElementDiscardingTerm :: Err m => Pos -> ByteString -> m (LocalIds, Type)

decodeTermElementDiscardingType :: Err m => Pos -> ByteString -> m (LocalIds, Term)

decodeDeclElement :: Err m => Word64 -> ByteString -> m (LocalIds, Decl Symbol)

legacy conversion helpers

getCycleLen :: EDB m => Hash -> m Word64

getDeclTypeByReference :: EDB m => Id -> m DeclType

componentByObjectId :: EDB m => ObjectId -> m [Id]

Codebase operations

Saving & loading terms

saveTermComponent :: EDB m => Hash -> [(Term Symbol, Type Symbol)] -> m ObjectId

saveReferenceH :: DB m => Reference' Text Hash -> m (Reference' TextId HashId)

Save the text and hash parts of a Reference to the database and substitute their ids.

c2xTerm :: forall m t d. Monad m => (Text -> m t) -> (Hash -> m d) -> Term Symbol -> Maybe (Type Symbol) -> m (LocalIds' t d, Term, Maybe Type)

implementation detail of c2{s,w}Term The Type is optional, because we don't store them for watch expression results.

loadTermWithTypeByReference :: EDB m => Id -> MaybeT m (Term Symbol, Type Symbol)

loadTermByReference :: EDB m => Id -> MaybeT m (Term Symbol)

loadTypeOfTermByTermReference :: EDB m => Id -> MaybeT m (Type Symbol)

s2cTermWithType :: EDB m => LocalIds -> Term -> Type -> m (Term Symbol, Type Symbol)

s2cTerm :: EDB m => LocalIds -> Term -> m (Term Symbol)

s2cTypeOfTerm :: EDB m => LocalIds -> Type -> m (Type Symbol)

localIdsToLookups :: Monad m => (t -> m Text) -> (d -> m Hash) -> LocalIds' t d -> m (LocalTextId -> Text, LocalDefnId -> Hash)

implementation detail of {s,w}2c*Term*

x2cTerm :: (LocalTextId -> Text) -> (LocalDefnId -> Hash) -> Term -> Term Symbol

implementation detail of {s,w}2c*Term*

x2cTType :: (LocalTextId -> Text) -> (LocalDefnId -> Hash) -> Type -> Type Symbol

implementation detail of {s,w}2c*Term*

lookupText :: forall m s w t. (MonadState s m, MonadWriter w m, Field1' s (Map t LocalTextId), Field1' w (Seq t), Ord t) => t -> m LocalTextId

lookupDefn :: forall m s w d. (MonadState s m, MonadWriter w m, Field2' s (Map d LocalDefnId), Field2' w (Seq d), Ord d) => d -> m LocalDefnId

lookup_ :: (MonadState s m, MonadWriter w m, Ord t) => Lens' s (Map t t') -> Lens' w (Seq t) -> (Word64 -> t') -> t -> m t'

shared implementation of lookupTextHelper and lookupDefnHelper Look up a value in the LUT, or append it.

c2sTerm :: EDB m => Term Symbol -> Type Symbol -> m (LocalIds, Term, Type)

Watch expressions

listWatches :: EDB m => WatchKind -> m [Id]

loadWatch :: EDB m => WatchKind -> Id -> MaybeT m (Term Symbol)

returns Nothing if the expression isn't cached.

saveWatch :: EDB m => WatchKind -> Id -> Term Symbol -> m ()

clearWatches :: DB m => m ()

c2wTerm :: EDB m => Term Symbol -> m (WatchLocalIds, Term)

w2cTerm :: EDB m => WatchLocalIds -> Term -> m (Term Symbol)

Saving & loading type decls

saveDeclComponent :: EDB m => Hash -> [Decl Symbol] -> m ObjectId

c2sDecl :: forall m t d. EDB m => (Text -> m t) -> (Hash -> m d) -> Decl Symbol -> m (LocalIds' t d, Decl Symbol)

loadDeclByReference :: EDB m => Id -> MaybeT m (Decl Symbol)

Branch transformation

s2cBranch :: EDB m => DbBranch -> m (Branch m)

type BranchSavingState = (Map Text LocalTextId, Map Hash LocalDefnId, Map PatchObjectId LocalPatchObjectId, Map (BranchObjectId, CausalHashId) LocalBranchChildId)

type BranchSavingWriter = (Seq Text, Seq Hash, Seq PatchObjectId, Seq (BranchObjectId, CausalHashId))

type BranchSavingConstraint m = (MonadState BranchSavingState m, MonadWriter BranchSavingWriter m)

type BranchSavingMonad m = StateT BranchSavingState (WriterT BranchSavingWriter m)

saveRootBranch :: EDB m => Causal m -> m (BranchObjectId, CausalHashId)

saveBranch :: EDB m => Causal m -> m (BranchObjectId, CausalHashId)

loadRootCausal :: EDB m => m (Causal m)

dataVersion :: DB m => m DataVersion

loadCausalBranchByCausalHash :: EDB m => CausalHash -> m (Maybe (Causal m))

loadCausalByCausalHashId :: EDB m => CausalHashId -> m (Causal m)

loadBranchByCausalHashId :: EDB m => CausalHashId -> m (Maybe (Branch m))

is this even a thing? loading a branch by causal hash? yes I guess so.

loadBranchByObjectId :: EDB m => BranchObjectId -> m (Branch m)

Patch transformation

loadPatchById :: EDB m => PatchObjectId -> m Patch

savePatch :: EDB m => PatchHash -> Patch -> m PatchObjectId

s2cPatch :: EDB m => Patch -> m Patch

deserializePatchObject :: EDB m => PatchObjectId -> m PatchFormat

lca :: EDB m => CausalHash -> CausalHash -> Connection -> Connection -> m (Maybe CausalHash)

before :: DB m => CausalHash -> CausalHash -> m (Maybe Bool)

Searches

termsHavingType :: EDB m => Reference -> m (Set Id)

typeReferenceForTerm :: EDB m => Id -> m ReferenceH

termsMentioningType :: EDB m => Reference -> m (Set Id)

addTypeToIndexForTerm :: EDB m => Id -> Reference -> m ()

addTypeMentionsToIndexForTerm :: EDB m => Id -> Set Reference -> m ()

componentReferencesByPrefix :: EDB m => ObjectType -> Text -> Maybe Pos -> m [Id]

termReferencesByPrefix :: EDB m => Text -> Maybe Word64 -> m [Id]

declReferencesByPrefix :: EDB m => Text -> Maybe Word64 -> m [Id]

termReferentsByPrefix :: EDB m => Text -> Maybe Word64 -> m [Id]

declReferentsByPrefix :: EDB m => Text -> Maybe Pos -> Maybe ConstructorId -> m [(Hash, Pos, Word64, DeclType, [ConstructorId])]

branchHashesByPrefix :: EDB m => ShortBranchHash -> m (Set BranchHash)

causalHashesByPrefix :: EDB m => ShortBranchHash -> m (Set CausalHash)

dependents :: EDB m => Reference -> m (Set Id)

returns a list of known definitions referencing r

derivedDependencies :: EDB m => Id -> m (Set Id)

returns empty set for unknown inputs; doesn't distinguish between term and decl