U.Codebase.Sqlite.Queries

Description

Some naming conventions used in this module:

32: the base32 representation of a hash

expect: retrieve something that's known to exist

load: retrieve something that's not known to exist (so the return type is a Maybe, or another container that could be empty)

save: idempotent (on conflict do nothing) insert, and return the id of the thing (usually)

text table

saveText :: Text -> Transaction TextId

loadTextId :: Text -> Transaction (Maybe TextId)

expectTextId :: Text -> Transaction TextId

expectText :: TextId -> Transaction Text

expectTextCheck :: SqliteExceptionReason e => TextId -> (Text -> Either e a) -> Transaction a

hash table

saveHash :: Base32Hex -> Transaction HashId

saveHashHash :: Hash -> Transaction HashId

loadHashId :: Base32Hex -> Transaction (Maybe HashId)

expectHash :: HashId -> Transaction Hash

expectHash32 :: HashId -> Transaction Base32Hex

loadHashIdByHash :: Hash -> Transaction (Maybe HashId)

expectHashIdByHash :: Hash -> Transaction HashId

saveCausalHash :: CausalHash -> Transaction CausalHashId

expectCausalHash :: CausalHashId -> Transaction CausalHash

saveBranchHash :: BranchHash -> Transaction BranchHashId

hash_object table

saveHashObject :: HashId -> ObjectId -> HashVersion -> Transaction ()

expectHashIdsForObject :: ObjectId -> Transaction (NonEmpty HashId)

hashIdWithVersionForObject :: ObjectId -> Transaction [(HashId, HashVersion)]

loadObjectIdForPrimaryHashId :: HashId -> Transaction (Maybe ObjectId)

expectObjectIdForPrimaryHashId :: HashId -> Transaction ObjectId

Not all hashes have corresponding objects; e.g., hashes of term types

loadObjectIdForPrimaryHash :: Hash -> Transaction (Maybe ObjectId)

expectObjectIdForPrimaryHash :: Hash -> Transaction ObjectId

loadPatchObjectIdForPrimaryHash :: PatchHash -> Transaction (Maybe PatchObjectId)

loadObjectIdForAnyHash :: Hash -> Transaction (Maybe ObjectId)

loadObjectIdForAnyHashId :: HashId -> Transaction (Maybe ObjectId)

expectObjectIdForAnyHashId :: HashId -> Transaction ObjectId

recordObjectRehash :: ObjectId -> ObjectId -> Transaction ()

recordObjectRehash old new records that object old was rehashed and inserted as a new object, new.

This function rewrites old's hash_object rows in place to point at the new object.

object table

saveObject :: HashId -> ObjectType -> ByteString -> Transaction ObjectId

expectObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

expectPrimaryHashByObjectId :: ObjectId -> Transaction Hash

All objects have corresponding hashes.

expectObjectWithHashIdAndType :: ObjectId -> Transaction (HashId, ObjectType, ByteString)

expectDeclObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a decl component object.

loadDeclObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a decl component object.

expectNamespaceObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a namespace object.

loadNamespaceObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a namespace object.

expectPatchObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a patch object.

loadPatchObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a patch object.

loadTermObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction (Maybe a)

Load a term component object.

expectTermObject :: SqliteExceptionReason e => ObjectId -> (ByteString -> Either e a) -> Transaction a

Expect a term component object.

namespace_root table

loadNamespaceRoot :: Transaction (Maybe CausalHashId)

setNamespaceRoot :: CausalHashId -> Transaction ()

expectNamespaceRoot :: Transaction CausalHashId

causals

causal table

saveCausal :: CausalHashId -> BranchHashId -> Transaction ()

Maybe we would generalize this to something other than NamespaceHash if we end up wanting to store other kinds of Causals here too.

isCausalHash :: HashId -> Transaction Bool

loadCausalHashIdByCausalHash :: CausalHash -> Transaction (Maybe CausalHashId)

expectCausalHashIdByCausalHash :: CausalHash -> Transaction CausalHashId

expectCausalValueHashId :: CausalHashId -> Transaction BranchHashId

]

loadCausalByCausalHash :: CausalHash -> Transaction (Maybe (CausalHashId, BranchHashId))

expectCausalByCausalHash :: CausalHash -> Transaction (CausalHashId, BranchHashId)

loadBranchObjectIdByCausalHashId :: CausalHashId -> Transaction (Maybe BranchObjectId)

expectBranchObjectIdByCausalHashId :: CausalHashId -> Transaction BranchObjectId

causal_parent table

saveCausalParents :: CausalHashId -> [CausalHashId] -> Transaction ()

loadCausalParents :: CausalHashId -> Transaction [CausalHashId]

before :: CausalHashId -> CausalHashId -> Transaction Bool

lca :: CausalHashId -> CausalHashId -> Transaction (Maybe CausalHashId)

watch table

saveWatch :: WatchKind -> IdH -> ByteString -> Transaction ()

loadWatch :: SqliteExceptionReason e => WatchKind -> IdH -> (ByteString -> Either e a) -> Transaction (Maybe a)

loadWatchesByWatchKind :: WatchKind -> Transaction [IdH]

loadWatchKindsByReference :: IdH -> Transaction [WatchKind]

clearWatches :: Transaction ()

indexes

dependents index

addToDependentsIndex :: Reference -> Id -> Transaction ()

getDependentsForDependency :: Reference -> Transaction [Id]

Get non-self, user-defined dependents of a dependency.

getDependentsForDependencyComponent :: ObjectId -> Transaction [Id]

getDependenciesForDependent :: Id -> Transaction [Reference]

Get non-self dependencies of a user-defined dependent.

getDependencyIdsForDependent :: Id -> Transaction [Id]

Get non-self, user-defined dependencies of a user-defined dependent.

migrations

countObjects :: Transaction Int

countCausals :: Transaction Int

countWatches :: Transaction Int

getCausalsWithoutBranchObjects :: Transaction [CausalHashId]

Finds all causals that refer to a branch for which we don't have an object stored. Although there are plans to support this in the future, currently all such cases are the result of database inconsistencies and are unexpected.

removeHashObjectsByHashingVersion :: HashVersion -> Transaction ()

Delete all hash objects of a given hash version. Leaves the corresponding hashes in the hash table alone.

type index

addToTypeIndex :: Reference' TextId HashId -> Id -> Transaction ()

getReferentsByType :: Reference' TextId HashId -> Transaction [Id]

getTypeReferenceForReferent :: Id -> Transaction (Reference' TextId HashId)

getTypeReferencesForComponent :: ObjectId -> Transaction [(Reference' TextId HashId, Id)]

type mentions index

addToTypeMentionsIndex :: Reference' TextId HashId -> Id -> Transaction ()

getReferentsByTypeMention :: Reference' TextId HashId -> Transaction [Id]

getTypeMentionsReferencesForComponent :: ObjectId -> Transaction [(Reference' TextId HashId, Id)]

hash prefix lookup

objectIdByBase32Prefix :: ObjectType -> Text -> Transaction [ObjectId]

namespaceHashIdByBase32Prefix :: Text -> Transaction [BranchHashId]

causalHashIdByBase32Prefix :: Text -> Transaction [CausalHashId]

garbage collection

garbageCollectObjectsWithoutHashes :: Transaction ()

Delete objects without hashes. An object typically *would* have a hash, but (for example) during a migration in which an object's hash may change, its corresponding hash_object row may be updated to point at a new version of that object. This procedure clears out all references to objects that do not have any corresponding hash_object rows.

garbageCollectWatchesWithoutObjects :: Transaction ()

Delete all

db misc

createSchema :: Transaction ()

schemaVersion :: Transaction SchemaVersion

expectSchemaVersion :: SchemaVersion -> Transaction ()

Expect the given schema version.

setSchemaVersion :: SchemaVersion -> Transaction ()

Orphan instances

FromField WatchKind

Instance details

Methods

fromField :: FieldParser WatchKind

FromField Base32Hex

Instance details

Methods

fromField :: FieldParser Base32Hex

ToField WatchKind

Instance details

Methods

toField :: WatchKind -> SQLData

ToField Base32Hex

Instance details

Methods

toField :: Base32Hex -> SQLData