U.Codebase.Sqlite.Queries

Constraint kinds

type DB m = (MonadIO m, MonadReader Connection m)

type EDB m = (DB m, Err m)

type Err m = (MonadError Integrity m, HasCallStack)

Error types

data Integrity

Constructors

UnknownHashId HashId

UnknownTextId TextId

UnknownObjectId ObjectId

UnknownCausalHashId CausalHashId

UnknownHash Hash

NoObjectForHashId HashId

NoObjectForPrimaryHashId HashId

NoNamespaceRoot

MultipleNamespaceRoots [CausalHashId]

NoSchemaVersion

MultipleSchemaVersions [SchemaVersion]

NoTypeIndexForTerm Id

text table

saveText :: DB m => Text -> m TextId

loadText :: DB m => Text -> m (Maybe TextId)

loadTextById :: EDB m => TextId -> m Text

hash table

saveHash :: DB m => Base32Hex -> m HashId

saveHashHash :: DB m => Hash -> m HashId

loadHashId :: DB m => Base32Hex -> m (Maybe HashId)

loadHashById :: EDB m => HashId -> m Base32Hex

loadHashHashById :: EDB m => HashId -> m Hash

loadHashIdByHash :: DB m => Hash -> m (Maybe HashId)

expectHashIdByHash :: EDB m => Hash -> m HashId

saveCausalHash :: DB m => CausalHash -> m CausalHashId

loadCausalHash :: EDB m => CausalHashId -> m CausalHash

saveBranchHash :: DB m => BranchHash -> m BranchHashId

hash_object table

saveHashObject :: DB m => HashId -> ObjectId -> Int -> m ()

hashIdsForObject :: DB m => ObjectId -> m (NonEmpty HashId)

hashIdWithVersionForObject :: DB m => ObjectId -> m [(HashId, Int)]

expectObjectIdForPrimaryHashId :: EDB m => HashId -> m ObjectId

Not all hashes have corresponding objects; e.g., hashes of term types

expectObjectIdForAnyHashId :: EDB m => HashId -> m ObjectId

maybeObjectIdForPrimaryHashId :: DB m => HashId -> m (Maybe ObjectId)

maybeObjectIdForAnyHashId :: DB m => HashId -> m (Maybe ObjectId)

recordObjectRehash :: DB m => ObjectId -> ObjectId -> m ()

recordObjectRehash old new records that object old was rehashed and inserted as a new object, new.

This function rewrites old's hash_object rows in place to point at the new object.

object table

saveObject :: DB m => HashId -> ObjectType -> ByteString -> m ObjectId

loadObjectById :: EDB m => ObjectId -> m ByteString

loadPrimaryHashByObjectId :: EDB m => ObjectId -> m Base32Hex

All objects have corresponding hashes.

loadObjectWithTypeById :: EDB m => ObjectId -> m (ObjectType, ByteString)

loadObjectWithHashIdAndTypeById :: EDB m => ObjectId -> m (HashId, ObjectType, ByteString)

updateObjectBlob :: DB m => ObjectId -> ByteString -> m ()

namespace_root table

loadMaybeNamespaceRoot :: EDB m => m (Maybe CausalHashId)

setNamespaceRoot :: forall m. DB m => CausalHashId -> m ()

loadNamespaceRoot :: EDB m => m CausalHashId

causals

causal table

saveCausal :: DB m => CausalHashId -> BranchHashId -> m ()

Maybe we would generalize this to something other than NamespaceHash if we end up wanting to store other kinds of Causals here too.

isCausalHash :: DB m => HashId -> m Bool

loadCausalHashIdByCausalHash :: DB m => CausalHash -> m (Maybe CausalHashId)

loadCausalValueHashId :: EDB m => CausalHashId -> m BranchHashId

]

loadCausalByCausalHash :: DB m => CausalHash -> m (Maybe (CausalHashId, BranchHashId))

loadBranchObjectIdByCausalHashId :: EDB m => CausalHashId -> m (Maybe BranchObjectId)

causal_parent table

saveCausalParents :: DB m => CausalHashId -> [CausalHashId] -> m ()

loadCausalParents :: DB m => CausalHashId -> m [CausalHashId]

before :: DB m => CausalHashId -> CausalHashId -> m Bool

lca :: CausalHashId -> CausalHashId -> Connection -> Connection -> IO (Maybe CausalHashId)

watch table

saveWatch :: DB m => WatchKind -> IdH -> ByteString -> m ()

loadWatch :: DB m => WatchKind -> IdH -> m (Maybe ByteString)

loadWatchesByWatchKind :: DB m => WatchKind -> m [IdH]

loadWatchKindsByReference :: DB m => IdH -> m [WatchKind]

clearWatches :: DB m => m ()

indexes

dependents index

addToDependentsIndex :: DB m => Reference -> Id -> m ()

getDependentsForDependency :: DB m => Reference -> m [Id]

Get non-self, user-defined dependents of a dependency.

getDependentsForDependencyComponent :: DB m => ObjectId -> m [Id]

getDependenciesForDependent :: DB m => Id -> m [Reference]

Get non-self dependencies of a user-defined dependent.

getDependencyIdsForDependent :: DB m => Id -> m [Id]

Get non-self, user-defined dependencies of a user-defined dependent.

type index

addToTypeIndex :: DB m => Reference' TextId HashId -> Id -> m ()

getReferentsByType :: DB m => Reference' TextId HashId -> m [Id]

getTypeReferenceForReferent :: EDB m => Id -> m (Reference' TextId HashId)

getTypeReferencesForComponent :: EDB m => ObjectId -> m [(Reference' TextId HashId, Id)]

type mentions index

addToTypeMentionsIndex :: DB m => Reference' TextId HashId -> Id -> m ()

getReferentsByTypeMention :: DB m => Reference' TextId HashId -> m [Id]

getTypeMentionsReferencesForComponent :: EDB m => ObjectId -> m [(Reference' TextId HashId, Id)]

hash prefix lookup

objectIdByBase32Prefix :: DB m => ObjectType -> Text -> m [ObjectId]

namespaceHashIdByBase32Prefix :: DB m => Text -> m [BranchHashId]

causalHashIdByBase32Prefix :: DB m => Text -> m [CausalHashId]

garbage collection

vacuum :: DB m => m ()

Clean the database and recover disk space. This is an expensive operation. Also note that it cannot be executed within a transaction.

garbageCollectObjectsWithoutHashes :: DB m => m ()

Delete objects without hashes. An object typically *would* have a hash, but (for example) during a migration in which an object's hash may change, its corresponding hash_object row may be updated to point at a new version of that object. This procedure clears out all references to objects that do not have any corresponding hash_object rows.

garbageCollectWatchesWithoutObjects :: DB m => m ()

Delete all

countObjects :: DB m => m Int

countCausals :: DB m => m Int

countWatches :: DB m => m Int

getCausalsWithoutBranchObjects :: DB m => m [CausalHashId]

Finds all causals that refer to a branch for which we don't have an object stored. Although there are plans to support this in the future, currently all such cases are the result of database inconsistencies and are unexpected.

db misc

createSchema :: (DB m, MonadUnliftIO m) => m ()

schemaVersion :: DB m => m SchemaVersion

setSchemaVersion :: DB m => SchemaVersion -> m ()

setFlags :: DB m => m ()

data DataVersion

The data version will increase if there has been any external modification to the database since the last observed data version.

dataVersion :: DB m => m DataVersion

savepoint :: DB m => String -> m ()

low-level transaction stuff

Create a savepoint, which is a named transaction which may wrap many nested sub-transactions.

release :: DB m => String -> m ()

Release a savepoint, which will commit the results once all wrapping transactions/savepoints are commited.

rollbackRelease :: DB m => String -> m ()

Roll back the savepoint and immediately release it. This effectively _aborts_ the savepoint, useful if an irrecoverable error is encountered.

rollbackTo :: DB m => String -> m ()

Roll the database back to its state from when the savepoint was created. Note: this also re-starts the savepoint and it must still be released if that is the intention. See rollbackRelease.

withSavepoint :: (MonadUnliftIO m, DB m) => String -> (m () -> m r) -> m r

Runs the provided action within a savepoint. Releases the savepoint on completion. If an exception occurs, the savepoint will be rolled-back and released, abandoning all changes.

withSavepoint_ :: (MonadUnliftIO m, DB m) => String -> m r -> m r

vacuumInto :: DB m => FilePath -> m ()

Copy the database into the specified location, performing a VACUUM in the process.

setJournalMode :: DB m => JournalMode -> m ()

traceConnectionFile :: DB m => m ()

print the active database filename

Orphan instances

FromField WatchKind

Instance details

Methods

fromField :: FieldParser WatchKind

FromField Base32Hex

Instance details

Methods

fromField :: FieldParser Base32Hex

ToField WatchKind

Instance details

Methods

toField :: WatchKind -> SQLData

ToField Base32Hex

Instance details

Methods

toField :: Base32Hex -> SQLData