U.Codebase.Sqlite.Queries

types

type DB m = (MonadIO m, MonadReader Connection m)

type EDB m = (DB m, Err m)

type Err m = (MonadError Integrity m, HasCallStack)

debugQuery :: Bool

debugThread :: Bool

debugConnection :: Bool

alwaysTraceOnCrash :: Bool

crashOnError :: Bool

throwError :: Err m => Integrity -> m c

data Integrity

Constructors

UnknownHashId HashId

UnknownTextId TextId

UnknownObjectId ObjectId

UnknownCausalHashId CausalHashId

UnknownHash Hash

NoObjectForHashId HashId

NoObjectForPrimaryHashId HashId

NoNamespaceRoot

MultipleNamespaceRoots [CausalHashId]

NoSchemaVersion

MultipleSchemaVersions [SchemaVersion]

NoTypeIndexForTerm Id

orError :: Err m => Integrity -> Maybe b -> m b

type TypeHashReference = Reference' TextId HashId

main squeeze

createSchema :: (DB m, MonadUnliftIO m) => m ()

setJournalMode :: DB m => JournalMode -> m ()

setFlags :: DB m => m ()

schemaVersion :: DB m => m SchemaVersion

saveHash :: DB m => Base32Hex -> m HashId

saveHashHash :: DB m => Hash -> m HashId

loadHashId :: DB m => Base32Hex -> m (Maybe HashId)

loadHashIdByHash :: DB m => Hash -> m (Maybe HashId)

saveCausalHash :: DB m => CausalHash -> m CausalHashId

saveBranchHash :: DB m => BranchHash -> m BranchHashId

loadCausalHashIdByCausalHash :: DB m => CausalHash -> m (Maybe CausalHashId)

loadCausalByCausalHash :: DB m => CausalHash -> m (Maybe (CausalHashId, BranchHashId))

expectHashIdByHash :: EDB m => Hash -> m HashId

loadHashHashById :: EDB m => HashId -> m Hash

loadHashById :: EDB m => HashId -> m Base32Hex

saveText :: DB m => Text -> m TextId

loadText :: DB m => Text -> m (Maybe TextId)

loadTextById :: EDB m => TextId -> m Text

saveHashObject :: DB m => HashId -> ObjectId -> Int -> m ()

saveObject :: DB m => HashId -> ObjectType -> ByteString -> m ObjectId

loadObjectById :: EDB m => ObjectId -> m ByteString

loadObjectWithTypeById :: EDB m => ObjectId -> m (ObjectType, ByteString)

loadObjectWithHashIdAndTypeById :: EDB m => ObjectId -> m (HashId, ObjectType, ByteString)

expectObjectIdForPrimaryHashId :: EDB m => HashId -> m ObjectId

Not all hashes have corresponding objects; e.g., hashes of term types

maybeObjectIdForPrimaryHashId :: DB m => HashId -> m (Maybe ObjectId)

expectObjectIdForAnyHashId :: EDB m => HashId -> m ObjectId

maybeObjectIdForAnyHashId :: DB m => HashId -> m (Maybe ObjectId)

loadPrimaryHashByObjectId :: EDB m => ObjectId -> m Base32Hex

All objects have corresponding hashes.

hashIdsForObject :: DB m => ObjectId -> m (NonEmpty HashId)

hashIdWithVersionForObject :: DB m => ObjectId -> m [(HashId, Int)]

updateObjectBlob :: DB m => ObjectId -> ByteString -> m ()

saveCausal :: DB m => CausalHashId -> BranchHashId -> m ()

Maybe we would generalize this to something other than NamespaceHash if we end up wanting to store other kinds of Causals here too.

loadCausalValueHashId :: EDB m => CausalHashId -> m BranchHashId

]

loadCausalHash :: EDB m => CausalHashId -> m CausalHash

loadMaybeCausalValueHashId :: DB m => HashId -> m (Maybe BranchHashId)

isCausalHash :: DB m => HashId -> m Bool

loadBranchObjectIdByCausalHashId :: EDB m => CausalHashId -> m (Maybe BranchObjectId)

saveCausalParents :: DB m => CausalHashId -> [CausalHashId] -> m ()

loadCausalParents :: DB m => CausalHashId -> m [CausalHashId]

newtype DataVersion

Constructors

DataVersion Int

dataVersion :: DB m => m DataVersion

loadNamespaceRoot :: EDB m => m CausalHashId

loadMaybeNamespaceRoot :: EDB m => m (Maybe CausalHashId)

setNamespaceRoot :: forall m. DB m => CausalHashId -> m ()

saveWatch :: DB m => WatchKind -> IdH -> ByteString -> m ()

loadWatch :: DB m => WatchKind -> IdH -> m (Maybe ByteString)

loadWatchKindsByReference :: DB m => IdH -> m [WatchKind]

loadWatchesByWatchKind :: DB m => WatchKind -> m [IdH]

clearWatches :: DB m => m ()

Index-building

addToTypeIndex :: DB m => Reference' TextId HashId -> Id -> m ()

getReferentsByType :: DB m => Reference' TextId HashId -> m [Id]

getTypeReferenceForReferent :: EDB m => Id -> m (Reference' TextId HashId)

getTypeReferencesForComponent :: EDB m => ObjectId -> m [(Reference' TextId HashId, Id)]

addToTypeMentionsIndex :: DB m => Reference' TextId HashId -> Id -> m ()

getReferentsByTypeMention :: DB m => Reference' TextId HashId -> m [Id]

getTypeMentionsReferencesForComponent :: EDB m => ObjectId -> m [(Reference' TextId HashId, Id)]

fixupTypeIndexRow :: (Reference' TextId HashId :. Id) -> (Reference' TextId HashId, Id)

addToDependentsIndex :: DB m => Reference -> Id -> m ()

getDependentsForDependency :: DB m => Reference -> m [Id]

getDependenciesForDependent :: DB m => Id -> m [Reference]

getDependencyIdsForDependent :: DB m => Id -> m [Id]

objectIdByBase32Prefix :: DB m => ObjectType -> Text -> m [ObjectId]

causalHashIdByBase32Prefix :: DB m => Text -> m [CausalHashId]

namespaceHashIdByBase32Prefix :: DB m => Text -> m [BranchHashId]

before :: DB m => CausalHashId -> CausalHashId -> m Bool

lca :: CausalHashId -> CausalHashId -> Connection -> Connection -> IO (Maybe CausalHashId)

ancestorSql :: String

helper functions

queryAtoms :: (DB f, ToRow q, FromField b, Show q, Show b) => Query -> q -> f [b]

composite input, atomic List output

queryAtoms_ :: (DB f, FromField b, Show b) => Query -> f [b]

no input, atomic List output

queryMaybe :: (DB f, ToRow q, FromRow b, Show q, Show b) => Query -> q -> f (Maybe b)

composite input, composite Maybe output

queryAtom :: (DB f, ToRow q, FromField b, Show q, Show b) => Query -> q -> f (Maybe b)

composite input, atomic Maybe output

queryOne :: Functor f => f (Maybe b) -> f b

Just output

query :: (DB m, ToRow q, FromRow r, Show q, Show r) => Query -> q -> m [r]

composite input, composite List output

query_ :: (DB m, FromRow r, Show r) => Query -> m [r]

no input, composite List output

debugHeader :: DB m => m String

queryTrace :: (MonadUnliftIO m, Show q, Show a) => String -> Query -> q -> m a -> m a

queryTrace_ :: (MonadUnliftIO m, Show a) => String -> Query -> m a -> m a

traceConnectionFile :: DB m => m ()

execute :: (DB m, ToRow q, Show q) => Query -> q -> m ()

execute_ :: DB m => Query -> m ()

executeMany :: (DB m, ToRow q, Show q) => Query -> [q] -> m ()

withImmediateTransaction :: (DB m, MonadUnliftIO m) => m a -> m a

transaction that blocks

savepoint :: DB m => String -> m ()

low-level transaction stuff

release :: DB m => String -> m ()

low-level transaction stuff

rollbackTo :: DB m => String -> m ()

low-level transaction stuff

rollbackRelease :: DB m => String -> m ()

low-level transaction stuff

orphan instances

Orphan instances

FromField WatchKind

Instance details

Methods

fromField :: FieldParser WatchKind

FromField Base32Hex

Instance details

Methods

fromField :: FieldParser Base32Hex

ToField WatchKind

Instance details

Methods

toField :: WatchKind -> SQLData

ToField Base32Hex

Instance details

Methods

toField :: Base32Hex -> SQLData