U.Codebase.Sqlite.Serialization

Documentation

decomposeBranchFormat :: MonadGet m => m SyncBranchFormat

decomposeDeclFormat :: MonadGet m => m SyncDeclFormat

decomposePatchFormat :: MonadGet m => m SyncPatchFormat

decomposeTermFormat :: MonadGet m => m SyncTermFormat

decomposeWatchFormat :: MonadGet m => m SyncWatchResultFormat

getBranchFormat :: MonadGet m => m BranchFormat

getLocalBranch :: MonadGet m => m LocalBranch

getDeclElement :: MonadGet m => m (Decl Symbol)

getDeclElementNumConstructors :: MonadGet m => m Int

Get the number of constructors in a decl element.

getDeclFormat :: MonadGet m => m DeclFormat

getPatchFormat :: MonadGet m => m PatchFormat

getLocalPatch :: MonadGet m => m LocalPatch

getTempCausalFormat :: MonadGet m => m TempCausalFormat

getTempDeclFormat :: MonadGet m => m TempDeclFormat

getTempNamespaceFormat :: MonadGet m => m TempNamespaceFormat

getTempPatchFormat :: MonadGet m => m TempPatchFormat

getTempTermFormat :: MonadGet m => m TempTermFormat

getTermAndType :: MonadGet m => m (Term, Type)

getTypeFromTermAndType :: MonadGet m => m Type

Decode ONLY the type of a term-component element. This is useful during sync and when we need the type of a term component element but don't want to decode the whole term (which can be expensive).

getTermFormat :: MonadGet m => m TermFormat

getWatchResultFormat :: MonadGet m => m WatchResultFormat

lookupDeclElement :: MonadGet m => Pos -> m (LocalIds, Decl Symbol)

lookupDeclElementNumConstructors :: MonadGet m => Pos -> m Int

lookupTermElement :: MonadGet m => Pos -> m (LocalIds, Term, Type)

lookupTermElementDiscardingTerm :: MonadGet m => Pos -> m (LocalIds, Type)

lookupTermElementDiscardingType :: MonadGet m => Pos -> m (LocalIds, Term)

putBranchFormat :: MonadPut m => BranchFormat -> m ()

putDeclFormat :: MonadPut m => DeclFormat -> m ()

putPatchFormat :: MonadPut m => PatchFormat -> m ()

putTempEntity :: MonadPut m => TempEntity -> m ()

putTermFormat :: MonadPut m => TermFormat -> m ()

putWatchResultFormat :: MonadPut m => WatchResultFormat -> m ()

recomposeBranchFormat :: MonadPut m => SyncBranchFormat -> m ()

recomposeDeclFormat :: MonadPut m => SyncDeclFormat -> m ()

recomposePatchFormat :: MonadPut m => SyncPatchFormat -> m ()

recomposeTermFormat :: MonadPut m => SyncTermFormat -> m ()

recomposeWatchFormat :: MonadPut m => SyncWatchResultFormat -> m ()

Exported for Share

putTermAndType :: MonadPut m => (Term, Type) -> m ()

putSingleTerm :: MonadPut m => Term -> m ()

Encode a single term without its type or component. Don't use this on its own unless you're encoding a watch result.

putDeclElement :: MonadPut m => DeclR TypeRef Symbol -> m ()

getSingleTerm :: MonadGet m => m Term

putLocalIdsWith :: MonadPut m => (t -> m ()) -> (d -> m ()) -> LocalIds' t d -> m ()

getLocalIdsWith :: MonadGet m => m t -> m d -> m (LocalIds' t d)

putLocalBranch :: MonadPut m => LocalBranch -> m ()

putLocalPatch :: MonadPut m => LocalPatch -> m ()